Switch syscall to syscalln (#13)
* Switch syscall to syscalln

This needs go at least 1.18, which is the first with syscalln.

If we want to get stuck on 1.13, we can still fix it by increasing nargs (they were all 1 smaller than needed to be), but I don't think it's worth it, as it's more error-prone, and go 1.13 is ancient.

Fixes #11

* Some formatting

Using newer go:build, plus putting external dependencies after built-in